backend/requirements.txt
# Backend Dependencies for AI Stock Risk Analysis Platform

# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Data processing
numpy>=1.24.0
pandas==2.2.3

# HTTP and API clients
httpx==0.27.2
requests==2.32.3

# AI and LLM
groq==0.11.0

# LangChain ecosystem
langchain==0.3.13
langchain-groq==0.2.1
langchain-community==0.3.13
duckduckgo-search==6.3.5

# News and data sources
yfinance==0.2.48
feedparser==6.0.11

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.12

# Logging and monitoring
loguru==0.7.2

# CORS
python-cors==1.0.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
